refactor(repositories): restructure the files in the repositories functions
